LoadClasses, MidLoad, Spawner
LoadClasses: Automates loading of Singletons now with ISingleton. Also
creates global data for singletons to be used when needed to be called.
MidLoad: Facilitiates the changing of level and changes to singleton if
needed in between levels.
Spanwer: Spawner now create object based on code. Command it to create
according to components.