Ability Manager Test
Error from Pixel - Will produce error if a save file has been loaded and if the edges has been covered with pixels, leading the a pixel unable to be produced and causes update to look for a pixel.

Error from LinkEdits - If a linked item is deleted, the source does not know and cannot remove the link. This led to errors.